Fix for #4855
This fix actually makes sense, but that it wasn't caught earlier
doesn't.  The construction of

   nodelet::Loader n(false)

was creating the first node handle and letting it go out of scope,
which was automagically calling ros::shutdown(), which is a dumb thing
for ros::NodeHandle to do automagically on destruction.

--HG--
extra : convert_revision : svn%3Aeb33c2ac-9c88-4c90-87e0-44a10359b0c3/stacks/common/trunk%4037219